requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
asyncpg
passlib
authlib
PyJWT
bcrypt
requests